Pacman (#7)
First version for pacman

+++ src/display/rasterdisplay.h
@@ -11,7 +11,7 @@ namespace ae
 		protected:
 			uint16_t _width;
 			uint16_t _height;
-			uint16_t* _pixels;
+			uint32_t* _pixels;
 			updatefn _callback;
 			SDL_Texture* _texture;
 
@@ -22,7 +22,7 @@ namespace ae
 			bool setSize(const uint16_t w, const uint16_t h) override;
 			bool registerCallback(updatefn) override;
 
-			bool setPixel(const uint16_t, const uint16_t, const uint16_t) override;
+			bool setPixel(const uint16_t, const uint16_t, const uint32_t) override;
 			bool init() override;
 			bool update(const SDL_Rect) override;
 		};
